Fix intersection of TVArrLike/TVecLike/TPArrLike with TInitUnc in HHBBC
Summary:
When we added the TVArrLike/TVecLike/TPArrLike (which
represent the union of ClsMeth and the types it can implicitly promote
to), we didn't add the types representing the static versions of
them. As a result, if you intersect them with TInitUnc (and other
types), you'll end up with an unrepresentable type, triggering
assertions.

Add TVArrLikeSA/TVecLikeSA/TPArrLikeSA to represent these types. I did
not use the typical TS* terminology because these do not represent
static types entirely (since ClsMeth may be counted). Instead the SA
means "static array" and indicates that the array part of the union is
only static arrays.

Add a bunch more tests to the type-system tests to check for these
issues. Also add tests testing the logic of ClsMeth being counted or
not depending on use_lowptr, which affects a lot of different
operations on the types.

While doing this, I noticed there's a similar issue with TStrLike. If
you intersect it with TPrim, you'll end up with TFunc | TCls, which is
not a representable type. Add a new type TFuncOrCls to represent this.

Reviewed By: markw65

Differential Revision: D20156897

fbshipit-source-id: 69c13cce68442ee73736a2527d2a0ffcb34def7b