Fixes for D20120480: Don't import Common, convert ref to RefCell
Summary:
The Common module just adds some functions to List and Option, so there is no Rust equivalent, and we can ignore the phrase `open Common` in hh_oxidize.

`std::cell::RefCell` is a suitable Rust representation of `ref` (we might want to use `Cell` in some circumstances, e.g. for `Copy` types, but always using `RefCell` is fine for now). This diff converts `ref` to `RefCell` in hh_oxidize and adds an implementation of OcamlRep which converts between the two.

Reviewed By: shiqicao

Differential Revision: D20131861

fbshipit-source-id: 63fe875755c8e7bc6913c756a78fd53c4654fb2a