Use glob imports and re-exports to simplify module-use logic
Summary:
I originally went out of my way to avoid glob imports and re-exports in hh_oxidize, in the interest of being explicit and avoiding unused imports (also because I believed at the time that Rust didn't support glob-exports in the way we needed). However, this makes it difficult to match OCaml semantics (of opens and includes, and also of plain module references, which do not require an import statement). This resulted in lots of logic in the Stringify module in hh_oxidize to implement those semantics while avoiding glob imports, glob re-exports, and unused imports.

I think this module is too clever for its own good. It would be much simpler if we allowed globs (to emulate open/include) and always imported every module in the crate root (to emulate the fact that modules can be referenced with no import statement).

Reviewed By: shiqicao

Differential Revision: D20131863

fbshipit-source-id: a1f35a21de41a6e680856ac4d32421ca5841323b