Derive Eq, Hash, Ord, PartialEq, and PartialOrd by default
Summary: We would like to compare and hash some of the types we oxidize, so it is easiest to derive these traits for all of them. All of these operations are already available on the OCaml side for all of our types (via polymorphic equals/hash/compare, or via `deriving eq`/`hash`/`ord`, or via handwritten comparison routines like `Typing_defs.ty_compare`). If we find that these capabilities are being misused--for example, if performance regressions are often being introduced by developers inserting large trees into hashed or ordered collections--we have the option of blacklisting the problematic types in hh_oxidize.

Differential Revision: D20144733

fbshipit-source-id: d708630c320c28e72b50b9fa83ef515e684a7ed2