Create `Naming_provider.add_type`
Summary:
Types in the reverse naming table are more complicated than constants and functions, since they could be (as of this writing) either a class, typedef, or record def. There's a lot more functions in the reverse naming table API, so I port them over one at a time.

There's an argument to be made that the API should offer fewer functions. For now, I'm going to include one function for each piece of functionality. This might be important when helping us decide if we want to back the reverse naming provider from the decl service, since it'll be easier to answer questions like "how often do we ask for information X about a symbol and throw all the rest away?" (which would tell us whether it's a good idea to calculate an entire decl that we then discard).

Reviewed By: ljw1004

Differential Revision: D20025029

fbshipit-source-id: 179d80e06e3166415549e2f688703b7b9ef6d672