Use an RWLock for static provenance
Summary:
Recursive provenance tagging is very slow. Profile data shows that a lot of the cost is due to provenance operations when we copy the initial array. When we copy an array, we initialize the new array's provenance with the old array's, instead of backtracing. This operation should be fast, but accessing provenance for static arrays is slow, because we use a single-exclusion mutex for the static array provenance map.

In this diff, we fix this problem by using a shared mutex instead. We can also add a helper to copy arrays without provenance, but that will take a bit more work.

Reviewed By: jjgriego

Differential Revision: D20160840

fbshipit-source-id: 8cdb50d92f8daa5a8b2c585741a684f9c7aa21e2