Incorporate lowptr mode into repo schema
Summary:
Since ClsMeth can be counted or uncounted depending on lowptr
mode, HHBBC will produce different bytecode depending on the
flag. Incorporate the lowptr mode into the repo schema to prevent
accidental usage of bytecode generated from a different mode.

This is only really an issue for HHBBC, so an alternate approach would
be encoding the lowptr mode into the global repo data after HHBBC
runs, and having HHVM refuse to load a repo which has a different
mode.

Reviewed By: mxw

Differential Revision: D20169592

fbshipit-source-id: adf152a97f27e9118c3799cb0f045bc88e2d90d9