Merge various ParamInfo flags into a single flags bitfield
Summary:
This is useful for two reasons.

One, allows inout flag to be also stored on the ParamInfo, without occupying
a full byte which we don't have. This will be used by the next diff that
simplifies the existing inout enforcement logic.

Two, provides a place for more bits that may be needed for linearity
enforcement.

Reviewed By: oulgen

Differential Revision: D20165686

fbshipit-source-id: 2fbf66258d563469faf0c2ce30775e53bd6d6744