Restore array RAT information
Summary:
The "vanilla" vs. "bespoke" array-like layouts are only visible to HHVM, so in general, we don't want to assume that HHBBC knows anything about them.

However, right now, we're overly pessimizing HHBBC types if the Eval.AllowBespokeArrayLikes flag is off. In this case, we can assume that any array-like is vanilla. The next diff in the stack achieves the same result, by inserting vanilla assertions everywhere that we might use an array-like's layout, but to avoid pessimizing current code, we fix typeFromRAT.

Note that "narrowToVanilla" doesn't affect non-array-like components of the type.

Reviewed By: ricklavoie

Differential Revision: D20179370

fbshipit-source-id: 051008c2faeb78b59751d18214b4c690bf81c865