ide_perf shows ready
Summary:
For ide_perf.sh, I want to see some definitive indication of when the process has initialized and has finished processing all files.

The problem was that sIDE would do initialization, then transition into the "Ready" state; then if there were files to process it would send "Done processing" eventually but if there were no files then it wouldn't send anything.

With this diff, we now send an LSP telemetry message after initialization has done and this message either says "0 files to process" (in which case we know everything's ready) or it says "N files to process" (in which case we know to wait for Done Processing)

Reviewed By: arxanas

Differential Revision: D19907193

fbshipit-source-id: 5945f2c2bffedb80e23941ea5e938410c305b94f