Get telemetry for all sharedmem+local heaps
Summary:
This introduces a function `Shared_mem.get_telemetry` which returns information on all heaps.

Some heaps have (1) some things stored in sharedmem, (2) a process-local stack of changes, (3) a process-local cache of values. Other heaps have only some of these three components.

This diff getss information for all components of all heaps.

It seems a bit costly, ~200ms, to compute all this data (because of the reachable_words). So this function doesn't yet hook it up into anything. I'll leave that for later.

Reviewed By: arxanas

Differential Revision: D19925288

fbshipit-source-id: 3706489e5dec140a3817e09ff2e21bdd8471c1bb