normalize heap names
Summary:
We have a lot of heaps! I found it hard, when reviewing their names, to tell at-a-glance which heap corresponded to which subsystem. This diff merely renames a few of shared heaps so they all have a meaningful prefix.

I verified that the description field is only ever used for logging - AI has a couple of "failwith" that mention it, and SharedMem.ml prints it too.

Also, for the compound heap "SharedMem.WithCache" which is composed of (1) NoCache which has sharedmem and stack, (2) LocalCache, I gave LocalCache the name "<description>:cache". That way it'll be easier to distinguish that it's just the local cache of a sharedmem heap; not a proper local cache in its own right.

Reviewed By: arxanas

Differential Revision: D19924994

fbshipit-source-id: c254e4f8547ab06c145e05ca324ee1e9f3aed7f5