Fix 7Up pretty printer for cores
Summary:
The 7Up pretty printer depended on materializing a temporary TV, but this only worked with a running process, and failed when debugging a core.

I considered adding a global "scratch" tv to hhvm that we could set for this situation, but didn't really like that idea. Instead, split out a pretty-print helper for tvs, and use it in both the object property printer, and the tv pretty printer.

Reviewed By: alexeyt

Differential Revision: D20178517

fbshipit-source-id: 470c43dc77f3bb2f3bd10ed2dae5043f0e4d519f