Guard locations on vanilla types
Summary:
This diff drops the "narrow the input to vanilla" logic from the JIT type-checking code.

Making this change causes many failures in our current irgen, because we never check for vanilla-ness anywhere. We don't want to pepper irgen with vanilla checks everywhere, though, for two reasons:
1. Doing so would make the irgen code ugly, spreading this vanilla requirement around several dozen files.
2. We want the vanilla requirement to be reflected in the guards for any region that depends on it, because that's how we will convert those regions to depend on "bespoke" layouts instead.

Instead of making each irgen function do these checks, we concentrate knowledge about which bytecodes need vanilla inputs in a new file: irgen-bespoke.h. The core logic in this file is in "getVanillaInputs", which takes an HHBC op and returns a list of guardable locations for which the existing irgen code assumes that the inputs are vanilla array-likes. We use this function to implement two public methods, "checkBespokeInputs" and "handleBespokeInputs". "checkBespokeInputs" breaks a tracelet if any inputs to an HHBC op are not known to be vanilla, and "handleBespokeInputs" actually adds a guard constrain on this condition. Both methods are fairly trivial wrappers around the core logic.

By forcing all vanilla logic through this file, we can implement many policies for handling these bespoke types. In particular, once we modify the runtime to be able to handle these types, these two methods will behave differently in when profiling vs. optimizing. When profiling, the vanilla array-like guards that we emit at the beginning of the region will also let logging array-likes enter the region, while "handleBespokeInputs" will unwrap these logging array-likes, push them through the vanilla code, and then re-wrap them. When optimizing, the vanilla array-like guards will force logging array-likes to side-exit the region, and we'll emit faster code for the selected bespoke case (if any) in "handleBespokeInputs".

One other policy we can implement in this file is to simply assert that types are vanilla, rather than checking. We do so if AllowBespokeArrayLikes is false, allowing us to check in this code without incurring a regression. (It is a significant regression if we turn the flag on.)

Reviewed By: mxw

Differential Revision: D19819685

fbshipit-source-id: 84560ab0bee4142e1d42431eb377243811cd2d0b