Fix folding of array conversions w/r/t provenance
Summary:
We neglected to apply the tag for the current bytecode for these casts.
Additionally, this means we now really shouldn't constprop when the current
function is PSF since this would produce the wrong tag on the result array.

This fixes a popular source of arrays stored in the repo w/o a tag.

Differential Revision: D20150341

fbshipit-source-id: d2a49827d1aa71f19f8fcc5e34b7d1938975d37a