Delete stack protection in GetScalarArray
Summary:
This was added to make the test in test/slow/parser/memory_exhaust.hhas pass
but we now have some stack protection in the deserializer and this also
interacts badly with backtrace.cpp--w/ array provenance enabled it's possible
to get into a loop where GetScalarArray exceeds the stack limit, so we raise a
ExtendedException, which computs a backtrace, which calls GetScalarArray, stack
limit still exceeded, and so on.

For now I want to sidestep the issue by eliding the check in GetScalarArray.
This puts it in the same category as e.g. ObjectData::release: functions that
will consume native stack space proportional to the nesting of their inputs but
that lack stack protection.

Reviewed By: ricklavoie

Differential Revision: D20162748

fbshipit-source-id: 3592d2e3bce2047970957a0d6ffc31656aa48146