Add runtime location provenance tags
Summary:
This adsd two more kinds of provenance tags (RuntimeLocation and
RuntimeLocationPoison) and extends usage of a request-local override previously
used for wait handles to try to establish an invariant of array provenance,
namely, that whenever the runtime requests a tag, it finds one.

`RuntimeLocation` tags can be requested via `ARRPROV_HERE()` or by installing a
RAII-wrapper `ARRPROV_USE_RUNTIME_LOCATION()` in scope. This produces a tag
with the C++ file and line number where the macro is invoked.

`RuntimeLocationPoison` tags are like regular runtime location tags but
additionally trigger an instrumentation in `tagFromPC`: namely, we should never
call `tagFromPC` while a poison tag is installed in the override. This is
useful to prevent backtracing out of e.g. the JIT or autoloader and ensure that
we are producing reasonable tags in those parts of the runtime.

Differential Revision: D20130382

fbshipit-source-id: 50d4814b9b989164d9f035d17dc35c99afae5107