Set up variadic inlined live types
Summary:
For optimized translations, we build a CFG to construct inline regions; for live translations, we select a tracelet instead. When we set up the live types for an inlined tracelet, we rely on DV funclets to initialize optional parameters, but for a function that takes a varargs param, there's no DV funclet for nargs == numNonVariadicParams. As a result, we'll emit an Uninit for this param and construct some nonsensical types when selecting this tracelet.

This bug is mostly asymptomatic, because as soon as we've selected the callee region, we'll emit IR starting with beginInlining, which calls emitPrologueLocals to handle all variadic cases correctly. However, it shows up when we start to do non-trivial type-checks on the variadic param (e.g. when AllowBespokeArrayLikes is on).

This fix isn't great. An alternative fix would be in irgen-call, if "nargs == numNonVariadicParams && callee->hasVariadicParams()", we'll push the empty varray on the stack the same way we construct a new varray if there are variadic params. But the irgen-call code does a lot more than this one inlining usage, so I opted for a narrower fix.

Reviewed By: markw65

Differential Revision: D20186421

fbshipit-source-id: 64ba5a94063fd5f7999b7d304930ed509046fd2a