check for OOMs in libxml
Summary:
Check for OOMs when allocating relatively large chunks of memory inside libxml. This doesn't fail the allocation (and thus doesn't prevent the memory pressure caused by huge requests), but it will cause the request to OOM fatal after the offending native function returns to PHP (assuming, of course, the process survives the kernel OOM killer). The new behavior is guarded by a new runtime option, which can be reused later as we add more checks of this flavor.

Ideally we'd return nullptr in allocation functions to avoid such memory pressure, and libxml actually allows that. Yet multiple places in the hhvm extension need fixing if we do it that way. We may still consider doing it in a later diff.

Reviewed By: markw65

Differential Revision: D20179914

fbshipit-source-id: b7d71f846fc034e8957675857c4454bcbfaf79c2