Convert locl ty variants
Summary:
We use a GADT in OCaml to partition `ty_` variants into the subset relevant to the decl phase and the subset relevant to the local inference phase. There isn't a way to do this in Rust (you can approximate it by using type parameters to make some variants uninstantiable, but exhaustive matching still requires you to list the uninstantiable variants, I think). As a result, because we originally only needed to represent decl tys (for the decl service), I just filtered out the locl variants.

We're now porting the typechecker and would like to represent locl types. We could oxidize the two variant subsets as two enums (D20153848), but andrewjkennedy suggested that the distinction between decl and locl is not critical enough to bother with distinguishing them in this way. Instead, we can just oxidize the GADT as if it were a normal variant type.

Reviewed By: andrewjkennedy

Differential Revision: D20162805

fbshipit-source-id: 754c6b3ca5c00583cfeac99afaf5a6ed5e1edcbb