Simplify threading of JSONs in hh_single_compile
Summary:
hh_single_compile currently threads through a list of
elements of type `Hh_json.json option`, which is overly complicated
and does not match the interface of Options_ffi module.

Change it to `string`, and omit elements that are `None`;
this way it's straightforward to swap a call to:

  Hhbc_options.apply_config_overrides_statelessly

with a Rust code behind FFI that does the same thing:

  Options_ffi.from_configs

so that Emitter can be partially migrated to Rust.

Differential Revision: D20050105

fbshipit-source-id: 4d9d1f47c5a2a122be14d9fc1ae565347bc9c2f3