Add flag to treat partial mode as strict
Summary:
As we get ready to get rid of partial mode, we'll need a mechanism to migrate away from it - as well as allow OSS to remain partial until things are cleaned up.

This flag also effectively removes user-denotable file modes from the language since `// experimental` is unused (soon to be removed) and `// partial` just means strict when the flag is on. It's weird to report errors on what are effectively comments.

Differential Revision: D20104459

fbshipit-source-id: 549f25e8ecef180257c1363c5d430d5f5b023845