Suppress formatting for FIXMEs based on line ranges rather than nodes
Summary:
Some time ago, we began suppressing formatting for the node following an HH_FIXME, since reformatting such a node can move the FIXME'd expression away from the FIXME comment, causing a Hack error.

Because we rolled it out shortly after implementing the hackfmt-ignore comment (which suppresses formatting for the node following the comment), and our time was limited, we used the same logic to implement formatting-suppression for FIXMEs (i.e., when a FIXME occurs before a node, we do not format that node). Since we didn't want to suppress formatting for entire nodes when they are very large (like a class declaration), we used a heuristic to only suppress formatting for nodes which spanned fewer than 3 lines.

This heuristic is not good enough. There are still many cases where failing to suppress formatting causes Hack errors. Since the typechecker understands FIXMEs on a per-line basis (they suppress errors with the given code on the line they're on, and on the line following it), and hackfmt is quite good at formatting ranges of lines, we should be able to suppress formatting for FIXMEs in a way that is guaranteed not to cause Hack errors. We just need to take the range of bytes to be formatted in the file and subtract the ranges of the FIXME lines from it. We delayed actually implementing this for quite some time due to a focus on server performance and scalability, but it's causing enough problems that we should just take care of it now.

This diff does not implement fixme-formatting-suppression for node-formatting or as-you-type-formatting. This is partially because I think it might be more difficult than full-file, range, and diff-formatting (which are the modes for which formatting-suppression is implemented here). But it's also because these APIs are not used in use cases where fixme-formatting-suppression is critical. AFAICT, node-formatting is used only by BigCode for neural code search (where I don't think this feature is relevant). As-you-type-formatting is an IDE feature--if auto-reformatting a statement (after typing a semicolon) caused a Hack error in that statement, it might be annoying, but I think we can expect that the error will be noticed and fixed by the developer.

Reviewed By: arxanas

Differential Revision: D20124500

fbshipit-source-id: 4b6b93df0c031c2316ba5a334364f94637233d26