hh_shared returns number of bytes removed
Summary:
I want to get an idea of how many values are stored on the shared heaps, and what is their size.

One building block, to obtain a cheap always-on record of how many bytes are in a shared heap, is for the C code simply to return how many bytes got freed on a "remove" event. That's what this diff provides.

I'll leave it to a future diff to take this "removed" number and do something with it.

Reviewed By: zlandau

Differential Revision: D19925129

fbshipit-source-id: 6d4ae510ce680257fd6f73edc63fdd810e70ba20