implement local linearization provider
Summary:
This is the important diff which switches ServerlessIDE away from shared-mem linearization provider to a local-mem LRU provider.

I happened to make the LRU provider be a single LRU hashtbl which stores both partial and completed linearizations. It'd be possible to separate then into two separate LRUs and come up with measurements about how much space to apportion for each. But my general assumption is that it's better to have fewer LRUs so that space-apportioning happens inherently.

The underlying data-structure for partial linearizations is an `mro_element Sequence.t` where the lambda for computing next elements is a heavy-weight thing that fetches other linearizations and shallow decls as needed.

The underlying data-structure for complete linearizations is just a `mro_element list` with no lambda involved.

I decided to store them both as `mro_element Sequence.t` where the closure for completed linearizations is just the trivial one from Sequence.to_list. This made the code cleaner.

(it would instead have been possible to store the actual underlying `mro_element list` in the local linearization_cache, with a bit more code, and call Sequence.to_list for every consumer of it. That's what the sharedmem provider does, since it would be kind of sketchy to store closures in sharedmem. I mean, I guess it'd work, but it'd weird me out.)

Differential Revision: D20120230

fbshipit-source-id: 61898741c29100a6aa81a8ce39c068ead2fb60a2