move quarantine to linearization_provider
Summary:
This is a trivial diff. All it does is move some APIs out of Decl_linearize and into Linearization_provider.

(The idea is that Linearization_provider is the backing store for linearizations, and hence is the thing that has to be quarantined. Decl_linearize is merely one consumer of it. Decl_redecl and Decl_provider are other consumers of it.)

Differential Revision: D20120115

fbshipit-source-id: 6ebb30063e578802a9726d57ce1ae1124a84561f