a stub linearization_provider
Summary:
Linearizations are the last part of decls to port off shared memory.

waleed had written D19592287 but, based on the review feedback, I wanted to start again with a slightly different approach.

This small diff merely adds a stub for the linearization provider. It doesn't do anything yet.

(I had to move the type `linearization` out of the implementation of decl_linearize, into decl_defs, so it could be shared).

Differential Revision: D20119945

fbshipit-source-id: ea660c5320abbbd6568f969e959aa8012201684c