Tweak API of decl_linearize
Summary:
This is a tiny no-op diff which merely adds an extra (unused for now) parameter to a few APIs. It'll be used further down the stack.

(The end goal is to move linearization out of shared memory and into a provider. To do that, we'll need to pass ctx around).

Differential Revision: D20119973

fbshipit-source-id: 3d1ceda836637a3c0e9b45a32885e7f1edd8be74