Store parser options in `Provider_context.t`
Summary:
The parser options are fundamentally required in order to access decls, since you will likely have to parse files in order to get the structure of a given decl. Store them in the `Provider_context.t` alongside the typechecker options.

Note that both `ParserOptions.t` and `TypecheckerOptions.t` are aliases for the same type (`GlobalOptions.t`). However, their values may be different in practice in the running of the program, since they're constructed individually out of the server configuration in `.hhconfig`.

It's possible that we will want to make these two not aliases for the same type in the future, in which case it would be helpful to be disciplined and split them up in this way.

Reviewed By: jewelpit

Differential Revision: D19870430

fbshipit-source-id: fdcd00f1f0383ba7b13b422b4e6f6e615d39f604