Fix race condition in Package::parse
Summary:
There's a bit of a race condition involving Repo and RuntimeOptions in
Package::parse. When creating a Repo (which is per-thread), we'll read
the relevant RuntimeOptions for configeration. This is usually fine
because the RuntimeOptions do not change (at least not after very
early startup). However that's not the case in Package::parse. We
create a thread here (which creates a Repo), and then sometime after
creating the thread, we might then modify the RuntimeOptions that the
Repo constructor reads. This means that there's a chance the thread
will read the options after they've been subsequently modified. It's
not clear to me what the total effect of this is, but it does cause
assertions to fire (because the RuntimeOptions are changing and making
the logic inconsistent).

This happens quite frequently when running the repo tests with the
test runner (presumably thread contention makes the race more likely
to happen).

Fix this by using a folly::Baton to mark when the newly created thread
has initialized its copy of the Repo. The creating thread will wait on
the baton so we will not modify any of the RuntimeOptions until after
the created thread has read them.

Reviewed By: markw65

Differential Revision: D20197785

fbshipit-source-id: 51a4fa76d16796ae3a9302ee6e43f35f246e13e2