Use throwable_to_string on more callback exception logging locations
Summary:
We know this is technically unsafe wrt reactivity because `__toString` is not final on an exception. However, we know these to be problems and should separate them from the more general 'implicitly invoked toString on an object' case.

It's common to have empty exception messages, when the exception type itself is meaningful enough. Due to that, let's not assume the empty string means 'toString() failed'.

Reviewed By: alexeyt

Differential Revision: D20129459

fbshipit-source-id: f31af39b3eaef952ae42bb7c41dccbb1bcae95c5