Alternative localization of PU
Summary:
The current implementation of PU localization suffers from a type hole
since it doesn't update correctly the constaints on some postponed type
variable checks.

A fix for this is to make the implementation closer to how `Taccess` are
handled.

In this diff, we do the following modifications to the system:
- `locl_ty` no longer have "flexible" PU dependent types (`C:E:@#n:T`), only
  rigid one that can subtype with themselves, or non resolved PU types
  (`C:E:ExplicitAtomName:T`).
- During localization, "flexible" PU dependent types are "hidden away" using
  a type variable, in the same way `Taccess` are dealt with.
- Once a variable is solved, we update all the potential related PU dependent
  types constraint.

Reviewed By: CatherineGasnier

Differential Revision: D20116350

fbshipit-source-id: a6efb3a332d5def8fb9f38f1713abd450a46222f