Track free bytes held in slab alocators
Summary:
When looking at rss and memory usage, it can be interesting to know
the amount of free memory available for use by requests that is considered used
by the system.  Here we add a counter to slab managers that tracks the number
of bytes currently held by the slab manager.

Reviewed By: mxw

Differential Revision: D20198640

fbshipit-source-id: f5cdb0762d31065c48485531e9aefbf9dbc02c37