Expose JEMalloc dirty page count
Summary:
As mentioned in the comment inline, pages in JEMalloc have 1 of three
states:
- active: In use by the application
- dirty: Previously used by the application, currently unused but not returned
  to the system in case the application needs more pages.  This is cheaper than
  looping in the kernel via madvise
- muzzy: Previously used by the application and currently unused.  These pages
  differ from dirty pages in that the kernel has been notified that they could
  be reclaimed (via `madvise(MADV_FREE)`).  They are not truly returned to the
  system (like with `madvised(MADV_DONTNEED)`) becuase the kernel will only
  reclaim the pages if under pressure.  This is better because upon future use
  a page may already be in place and not require a ZFOD.

By default when pages become unused they spend 10s in the dirty state,
followed by 10s in the muzzy state, and finally are force returned to the
kernel via `madvise(MADV_DONTNEED)`.

In order to better understand memory usage is useful to know how much space is
held in the dirty state.  This is memory that could be reclaimed (as it is
unused), but is not able to be as JEMalloc has not informed the kernel.  Using
this counter, and information about current system pressure, we can force
purge this space when the system is under pressure.

Reviewed By: mxw

Differential Revision: D20198642

fbshipit-source-id: 2cc184860429ef7450a5b032059ba90205abd971