Remove lifetime from Allocator
Summary:
As noted in T61819545, the unconstrained lifetime on `Arena::new` allows the caller to choose an arbitrary lifetime which may be longer than the arena actually lives, allowing use-after-free of allocated values.

We can properly constrain the lifetime of the values by removing Arena's lifetime parameter and associating Values with the lifetime of the *borrow* of the Arena with which they were allocated.

There seem to be no existing callers exercising the use-after-free vulnerability.

Reviewed By: dtolnay

Differential Revision: D20174829

fbshipit-source-id: ef347b3c6aba57af7f5170f37c4bacb4742ecbe8