rename+comments inside SharedMem.ml
Summary:
I'm just adding comments and making a rename.

It used to say "All the cache are configured by a module of type ConfigType". That wasn't actually true. It was only the two hashtbl caches (FreqCache and OrderedCache) that were parameterized by it. Moreover, FreqCache and OrderedCache were only ever used as S1/S2 within a LocalCache.

I've updated comments, and renamed, accordingly.

Reviewed By: arxanas

Differential Revision: D20124453

fbshipit-source-id: 4b5aa6e8811a63d0b00e5c9a7fe70ed18e731cca