Mark BlockBuilder::as_mut_ptr safe
Summary: This function was originally marked unsafe at an early stage in development, when I did not understand `unsafe` as well as I do now. I had marked all Block/BlockBuilder/Value functions which accepted or returned pointers `unsafe`. This is not necessary--for instance, it is always possible to cast `&T` to `*const T` in safe code using `as` (e.g., `&value as *const _`). It is only *dereferencing* a raw pointer which requires `unsafe`.

Reviewed By: dtolnay

Differential Revision: D20202434

fbshipit-source-id: 17b223684b7dce575d23a92c15d9c97c0a146e33