Avoid cloning blocks in emit_statement.rs
Summary: When emitting the body block in for, foreach, while, and do loops, the OCaml version wraps the block into a `Stmt(pos, block)` and calls a helper on that `Stmt`. This wrapping step in Rust requires cloning of the block, which is super expensive. The helper doesn't need the pos parameter, but only computes instructions from the block, so get rid of the wrapper to avoid cloning. (corresponding change in OCaml: D20236751)

Reviewed By: shiqicao

Differential Revision: D20237137

fbshipit-source-id: 8ae75d9265ef04bedb9860299d2c524eeb59b277