Introduce alternate server job-queue dequeue limiter
Summary:
Introduces a mechanism to limit number of concurrent jobs to a
fixed value.  When `Server.LegacyBehavior = false`, the job queue simply limits
active workers to the set MaxActiveWorkers value.  When
`Server.LegacyBehavior = true` the job queue behaves as it does today with various release
ratelimiting mechanisms at various thresholds.

Regardless of the `Server.LegacyBehavior` setting, if a thread is not allowed
to dequeue, it flushes its stack and such to free up resources.

Reviewed By: binliu19

Differential Revision: D20226679

fbshipit-source-id: b1b11a08c7ed73e3c167a8cfe23d53d47c802dc0