Whitelist layout-agnostic ops
Summary:
Rather than marking the ~150 ops that require vanilla inputs, mark the ~10 ops that can tolerate arbitrary array inputs.

Sorry for the back and forth on this choice. The main reason to mark ops with the vanilla requirement is to be explicit; the main reason to prefer this approach is to make development easier while we haven't enabled bespoke array-like checks at runtime. For now, I think it's better to assume ops have a vanilla requirement unless otherwise stated - it's easy for people to miss the requirement if they're working on other projects.

This approach is also easier to strip out from the array or to ignore based on flags at runtime - if we need to excise the vanilla types, we'll just eliminate the two calls to irgen-bespoke and make "checkLayoutFlag" trivial in the two places where we check it.

Reviewed By: ricklavoie

Differential Revision: D20193408

fbshipit-source-id: bc55ddb83679ddf523d384e8da9250f584f563f7