Get rid of unnecessary Stmt wrappers around Block in emit_statement
Summary: When emitting the body block in for, foreach, while, and do loops, the OCaml version wraps the block into a Stmt(pos, block) and calls a helper on that Stmt. This wrapping step, when ported to Rust, requires cloning of the block, which is super expensive. The helper doesn't need the pos parameter, but only computes instructions from the block, so get rid of the wrapper to avoid cloning.

Differential Revision: D20236751

fbshipit-source-id: c95b0f9ac09b278717114fea700f2ed734dbe812