Delete Typing_get_locals
Summary:
Looking at cleaning up naming.ml and was a bit intrigued by the functorization of naming.

It turns out that the majority of `Typing_get_locals` is dead/unreferenced. And the only portion of `Typing_get_locals` that is called (`lvalue`) is relatively contained and does not need to be different on the decl path and the typing path. Thus, we can get rid of the modularization of Naming - delete `Typing_get_locals` and `Typing_naming_body`.

Reviewed By: andrewjkennedy, Wilfred

Differential Revision: D20239664

fbshipit-source-id: 3d8ab3e004258ab2f86626514bd1ebdaf1d2bcc9