Don't permit the VM to call back into the debugger while executing client commands
Summary: To fix: S197024. Commands that can potentially invoke Hack code on the request must not be allowed to break back into the debugger. Since the program state is paused, calling back into the debugger hook causes the thread to enter the debugger command queue and suspend, with nothing to ever wake it again causing all request threads to hang forever.

Reviewed By: alexeyt

Differential Revision: D20277511

fbshipit-source-id: 7da41c2fc3d7d0c8057e26b7fdbe42eed1c473e6