Fix lint for gdb python files
Summary:
I've made a few updates to the gdb python files recently, and every time
my diffs are spammed with unrelated lint errors.

I don't like some of these changes, but it seems worth it to get clean
diffs.

I've not tried to fix the "function too complex" warnings, because I
want this to be a functional no-op

Reviewed By: oulgen, alexeyt

Differential Revision: D20260820

fbshipit-source-id: 076d75969ac436ead25823ed5e0d312a7bbc53a1