Precompute static enum values
Summary:
In most cases, we can compute a static array for an enum's getNames or getValues methods. Burn these values into the TC and improve the code around these calls.

I'll try optimizing coerce and isValid as well as a follow-up. The only annoying complication is intish-cast, but I think I can avoid it it most cases.

Reviewed By: paulbiss

Differential Revision: D20113013

fbshipit-source-id: fef1eb7cd6127d79374bf71fa5ac681d2cdc2d11