Get rid of unnecessary Stmt wrappers in emit_statement
Summary:
Similar to D20236751

When emitting the body block in `using` and `try/catch` statements, the OCaml version wraps the block into a Stmt(pos, block) and calls a helper on that Stmt. This wrapping step, when ported to Rust, will require cloning of the block, which is super expensive. Get rid of the wrapper to avoid cloning.

Reviewed By: shiqicao

Differential Revision: D20311559

fbshipit-source-id: d4e5eaf5a382e5d28b602f210a7020c34476854e